requirements.txt
streamlit==1.31.1
Pillow==10.2.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
torch==2.0.1
torchvision==0.15.2
ultralytics>=8.0.0
manga-ocr>=0.1.8
python-dotenv>=1.0.0
deepl>=1.12.0
googletrans==4.0.0rc1
transformers>=4.30.0
matplotlib>=3.5.0